Seperate device policy from device module
