+++ features/admin.device-policies.v1/hooks/use-get-device-policy-by-id.ts
@@ -16,37 +16,34 @@
  * under the License.
  */
 
+import { HttpMethods } from "@wso2is/core/models";
 import { store } from "@wso2is/admin.core.v1/store";
 import useRequest, { RequestErrorInterface, RequestResultInterface } from "@wso2is/admin.core.v1/hooks/use-request";
-import { HttpMethods } from "@wso2is/core/models";
-import { DevicePolicyFieldDefinitionInterface, DevicePlatformType } from "../models/devices";
+import { DevicePolicyResponseInterface } from "../models/device-policy";
 
 /**
- * Hook to fetch device policy field metadata for a given platform.
+ * Hook to fetch a single device policy by ID.
  *
- * @param platform - Platform to filter metadata by.
+ * @param policyId - The policy ID to fetch.
  * @param shouldFetch - Whether to trigger the request.
- * @returns SWR result with field definitions.
+ * @returns SWR result with the device policy.
  */
-const useGetDevicePolicyMetadata = (
-    platform: DevicePlatformType | null,
+const useGetDevicePolicyById = (
+    policyId: string,
     shouldFetch: boolean = true
-): RequestResultInterface<DevicePolicyFieldDefinitionInterface[], RequestErrorInterface> => {
-    const baseUrl: string = store.getState().config.endpoints.devicePolicyMetadata;
-    const url: string = platform ? `${ baseUrl }?platform=${ platform }` : baseUrl;
-
+): RequestResultInterface<DevicePolicyResponseInterface, RequestErrorInterface> => {
     const requestConfig: { url: string; method: string; headers: Record<string, string> } = {
         headers: { "Content-Type": "application/json" },
         method: HttpMethods.GET,
-        url
+        url: `${ store.getState().config.endpoints.devicePolicies }/${ policyId }`
     };
 
     const { data, isLoading, isValidating, error, mutate } = useRequest<
-        DevicePolicyFieldDefinitionInterface[],
+        DevicePolicyResponseInterface,
         RequestErrorInterface
-    >(shouldFetch && platform ? requestConfig : null);
+    >(shouldFetch ? requestConfig : null);
 
     return { data, error, isLoading, isValidating, mutate };
 };
 
-export default useGetDevicePolicyMetadata;
+export default useGetDevicePolicyById;
