+++ features/admin.device-policies.v1/package.json
@@ -1,8 +1,8 @@
 {
     "private": true,
-    "name": "@wso2is/admin.devices.v1",
+    "name": "@wso2is/admin.device-policies.v1",
     "version": "1.0.0",
-    "description": "Device management UI for WSO2 Identity Server console",
+    "description": "Device assurance policies UI for WSO2 Identity Server console",
     "author": "WSO2",
     "license": "Apache-2.0",
     "main": "public-api.ts",
@@ -14,8 +14,10 @@
         "@mui/material": "catalog:",
         "@mui/system": "catalog:",
         "@oxygen-ui/react": "catalog:",
+        "@oxygen-ui/react-icons": "^2.4.6",
         "@wso2is/access-control": "^3.5.6",
         "@wso2is/admin.core.v1": "workspace:^",
+        "@wso2is/admin.rules.v1": "workspace:^",
         "@wso2is/core": "^2.13.0",
         "@wso2is/react-components": "^2.10.0",
         "i18next": "^21.9.1",
