+++ features/admin.device-policies.v1/tsconfig.json
@@ -1,0 +1,42 @@
+{
+  "compilerOptions": {
+      "jsx": "react",
+      "declaration": true,
+      "declarationDir": "dist/esm/types",
+      "lib": [ "ESNext", "DOM", "DOM.Iterable", "ScriptHost" ],
+      "resolveJsonModule": true,
+      "skipDefaultLibCheck": true,
+      "types": [ "node", "webpack-env", "jest", "@testing-library/jest-dom" ],
+      "emitDecoratorMetadata": true,
+      "esModuleInterop": true,
+      "experimentalDecorators": true,
+      "importHelpers": true,
+      "module": "esnext",
+      "moduleResolution": "node",
+      "skipLibCheck": true,
+      "sourceMap": true,
+      "target": "es2015"
+  },
+  "exclude": [
+      "build",
+      "cache",
+      "coverage",
+      "dist",
+      "node_modules",
+      "scripts",
+      "**/test-configs/*",
+      "jest.config.ts",
+      "**/tests/*",
+      "**/__tests__/*",
+      "**/__mocks__/*",
+      "**/*.test.js",
+      "**/*.test.jsx",
+      "**/*.test.ts",
+      "**/*.test.tsx",
+      "**/*.spec.js",
+      "**/*.spec.jsx",
+      "**/*.spec.ts",
+      "**/*.spec.tsx"
+  ],
+  "compileOnSave": false
+}
