+++ features/admin.device-policies.v1/types/imports.d.ts
@@ -16,7 +16,25 @@
  * under the License.
  */
 
-export interface DevicesResourceEndpointsInterface {
-    devices: string;
-    userDevices: string;
+declare module "*.json" {
+    const value: any;
+
+    export default value;
 }
+
+declare module "*.svg" {
+    import React = require("react");
+
+    export const ReactComponent: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
+    const src: string;
+
+    export default src;
+}
+
+declare module "*.png" {
+    const content: string;
+
+    export default content;
+}
+
+declare module "*.md";
