+++ features/admin.devices.v1/models/devices.ts
@@ -30,101 +30,3 @@ export interface DeviceResponseInterface {
 export interface DevicePatchRequestInterface {
     deviceName: string;
 }
-
-export interface PolicyExpressionInterface {
-    field: string;
-    operator: string;
-    value: string;
-}
-
-export interface PolicyRuleInterface {
-    condition: "AND" | "OR";
-    expressions: PolicyExpressionInterface[];
-}
-
-export interface PolicyResourceRequestInterface {
-    target: string;
-    resourceType?: "RULE" | "ACTION";
-    rule: PolicyRuleInterface;
-}
-
-export interface PolicyRequestInterface {
-    name: string;
-    resources: PolicyResourceRequestInterface[];
-}
-
-export interface DevicePolicyExpressionValueInterface {
-    type: string;
-    value: string;
-}
-
-export interface DevicePolicyExpressionInterface {
-    field: string;
-    displayName: string;
-    operator: string;
-    value: DevicePolicyExpressionValueInterface;
-}
-
-export interface DevicePolicyRuleGroupInterface {
-    expressions: DevicePolicyExpressionInterface[];
-}
-
-export interface DevicePolicyFullRuleInterface {
-    rules: DevicePolicyRuleGroupInterface[];
-}
-
-export interface PolicyResourceResponseInterface {
-    id: string;
-    target: string;
-    resourceType: "RULE" | "ACTION";
-    resourceId: string;
-    rule?: DevicePolicyFullRuleInterface;
-}
-
-export interface DevicePolicyResponseInterface {
-    id: string;
-    name: string;
-    resources?: PolicyResourceResponseInterface[];
-}
-
-export interface PolicyListItemInterface {
-    id: string;
-    name: string;
-    self?: string;
-}
-
-export interface PolicyListLinkInterface {
-    href: string;
-    rel: string;
-}
-
-export interface PolicyListResponseInterface {
-    totalResults: number;
-    startIndex: number;
-    count: number;
-    policies: PolicyListItemInterface[];
-    links?: PolicyListLinkInterface[];
-}
-
-export type DevicePlatformType = "android" | "ios" | "macos" | "windows";
-
-export interface DevicePolicyMetadataValueOptionInterface {
-    name: string;
-    displayName: string;
-}
-
-export interface DevicePolicyFieldDefinitionInterface {
-    field: {
-        name: string;
-        displayName: string;
-    };
-    operators: {
-        name: string;
-        displayName: string;
-    }[];
-    value: {
-        inputType: "INPUT" | "OPTIONS";
-        valueType: "STRING" | "NUMBER" | "BOOLEAN";
-        values?: DevicePolicyMetadataValueOptionInterface[];
-    };
-}
