+++ features/admin.devices.v1/public-api.ts
@@ -16,13 +16,8 @@
  * under the License.
  */
 
-export * from "./api/device-policies";
 export * from "./configs/endpoints";
-export * from "./hooks/use-get-device-policies";
-export * from "./hooks/use-get-device-policy-by-id";
-export * from "./hooks/use-get-device-policy-metadata";
 export { default as useGetDevicesByUserId } from "./hooks/use-get-devices-by-user-id";
 export * from "./models/devices";
-export * from "./pages/device-assurance-policies";
-export * from "./pages/device-policy-edit";
+export * from "./models/endpoints";
 export * from "./pages/devices";
