+++ features/admin.registration-flow-builder.v1/package.json
@@ -18,7 +18,7 @@
         "@wso2is/access-control": "^3.5.6",
         "@wso2is/admin.claims.v1": "^2.31.2",
         "@wso2is/admin.core.v1": "^2.57.6",
-        "@wso2is/admin.devices.v1": "workspace:^",
+        "@wso2is/admin.device-policies.v1": "workspace:^",
         "@wso2is/admin.feature-gate.v1": "^1.9.5",
         "@wso2is/admin.flow-builder-core.v1": "^0.10.212",
         "@wso2is/admin.flows.v1": "workspace:^",
